pyproject.toml
[project]
name = "pars-wb-t"
version = "0.1.0"
description = ""
authors = [
    {name = "Anton Novojilov",email = "wowlass63@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "ruff (>=0.14.10,<0.15.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
